Splitting Range and Repeat tests into their own files
