+++ src/System.Linq/tests/System.Linq.Tests.csproj
@@ -29,7 +29,8 @@
     <Compile Include="Performance\SkipTakePerf.cs" />
     <Compile Include="Performance\WhereSelectPerf.cs" />
     <Compile Include="Performance\ToArrayToListToDictionaryPerf.cs" />
-    <Compile Include="RangeRepeatTests.cs" />
+    <Compile Include="RangeTests.cs" />
+    <Compile Include="RepeatTests.cs" />
     <Compile Include="WhereTests.cs" />
   </ItemGroup>
   <ItemGroup>
